+++ turkle/tests/test_views.py
@@ -70,7 +70,7 @@ class TestAcceptTask(TestCase):
 
 class TestAcceptNextTask(TestCase):
     def setUp(self):
-        project = Project.objects.create(
+        self.project = Project.objects.create(
             login_required=False,
             html_template='<p>${foo}: ${bar}</p><textarea>',
             name='foo',
@@ -79,7 +79,7 @@ class TestAcceptNextTask(TestCase):
             filename='foo.csv',
             login_required=False,
             name='foo',
-            project=project,
+            project=self.project,
         )
 
         self.task = Task.objects.create(
@@ -167,6 +167,44 @@ class TestAcceptNextTask(TestCase):
                                       kwargs={'batch_id': self.batch.id}))
         self.assertEqual(response.status_code, 302)
         self.assertTrue('{}/assignment/'.format(task_two.id) in response['Location'])
+
+    def test_accept_next_task__deactivated_batch(self):
+        self.batch.active = False
+        self.batch.save()
+
+        User.objects.create_user('testuser', password='secret')
+        self.assertEqual(self.task.taskassignment_set.count(), 0)
+
+        client = django.test.Client()
+        client.login(username='testuser', password='secret')
+        response = client.get(reverse('accept_next_task',
+                                      kwargs={'batch_id': self.batch.id}))
+        self.assertEqual(response.status_code, 302)
+        self.assertEqual(response['Location'], reverse('index'))
+        self.assertEqual(self.task.taskassignment_set.count(), 0)
+        messages = list(get_messages(response.wsgi_request))
+        self.assertEqual(len(messages), 1)
+        self.assertEqual(str(messages[0]),
+                         'No more Tasks available for Batch {}'.format(self.batch.name))
+
+    def test_accept_next_task__deactivated_project(self):
+        self.project.active = False
+        self.project.save()
+
+        User.objects.create_user('testuser', password='secret')
+        self.assertEqual(self.task.taskassignment_set.count(), 0)
+
+        client = django.test.Client()
+        client.login(username='testuser', password='secret')
+        response = client.get(reverse('accept_next_task',
+                                      kwargs={'batch_id': self.batch.id}))
+        self.assertEqual(response.status_code, 302)
+        self.assertEqual(response['Location'], reverse('index'))
+        self.assertEqual(self.task.taskassignment_set.count(), 0)
+        messages = list(get_messages(response.wsgi_request))
+        self.assertEqual(len(messages), 1)
+        self.assertEqual(str(messages[0]),
+                         'No more Tasks available for Batch {}'.format(self.batch.name))
 
 
 class TestDownloadBatchCSV(TestCase):
@@ -369,11 +407,11 @@ class TestIndexAbandonedAssignments(TestCase):
     def setUp(self):
         self.user = User.objects.create_user('testuser', password='secret')
 
-        project = Project()
-        project.save()
-        batch = Batch(project=project)
-        batch.save()
-        self.task = Task(batch=batch)
+        self.project = Project()
+        self.project.save()
+        self.batch = Batch(project=self.project)
+        self.batch.save()
+        self.task = Task(batch=self.batch)
         self.task.save()
 
     def test_index_abandoned_assignment(self):
@@ -388,6 +426,40 @@ class TestIndexAbandonedAssignments(TestCase):
         response = client.get(reverse('index'))
         self.assertEqual(response.status_code, 200)
         self.assertTrue(b'You have abandoned' in response.content)
+
+    def test_index_abandoned_assignment_from_inactive_batch(self):
+        # Don't show abandoned tasks from inactive batches
+        TaskAssignment(
+            assigned_to=self.user,
+            completed=False,
+            task=self.task,
+        ).save()
+
+        self.project.active = False
+        self.project.save()
+
+        client = django.test.Client()
+        client.login(username='testuser', password='secret')
+        response = client.get(reverse('index'))
+        self.assertEqual(response.status_code, 200)
+        self.assertFalse(b'You have abandoned' in response.content)
+
+    def test_index_abandoned_assignment_from_inactive_project(self):
+        # Don't show abandoned tasks from inactive projects
+        TaskAssignment(
+            assigned_to=self.user,
+            completed=False,
+            task=self.task,
+        ).save()
+
+        self.batch.active = False
+        self.batch.save()
+
+        client = django.test.Client()
+        client.login(username='testuser', password='secret')
+        response = client.get(reverse('index'))
+        self.assertEqual(response.status_code, 200)
+        self.assertFalse(b'You have abandoned' in response.content)
 
     def test_index_no_abandoned_assignments(self):
         TaskAssignment(